Use timegm instead of mktime and take offset into account
libgit2 represents time in the git_time struct as seconds from epoch in
local time and includes an offset from UTC in this struct.  Previously,
the offset was retained but not taken into account when converting the
broken-down tm time to a calendar time time_t. This commit attempts to
fix this by converting the tm struct into seconds since epoch in UTC and
adding the timezone-specific offset from the commit on top of it to
calculate the correct local time.

Unfortunately, this change requires timegm which is not mandated by
POSIX but the function seems to be available no most systems and there
doesn't seem to be any easy way to do this with POSIX time.h functions.
The _BSD_SOURCE feature test macro has been added to reflect this
requirement.